// dpeweb/src/app/service/model/MenuModel.ts
export class MenuModel {
    MENUID: any;
    MENUNAME: any;
    MENULEVEL: any;
    MENULEVELNAME: any;
    MENUPARENT: any;

    PARENTNAME: any;
    MENUNO: any;
    MENUGROUP: any;
    MENUTYPE: any;
    MENUICON: any;

    MENUIMAGE: any;
    MENUCONTROL: any;
    MENUACTION: any;
    RELATEID: any;
    MENUSHOW: any;

    MENUPAGETYPE: any;
    MENUDROP: any;
    RELATEGROUP: any;

}